NP-1858 Move all action to one file
